* simplify server creation logic * rename www.ts to server.ts * move server.ts and db.ts to src directory
